+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1
     - Added Money.fromString() constructor.
+    - Added relational operators (<, <=, >, >=).
 ## 0.1.0+1
     - Fixes README.md
 ## 0.1.0